requirements.txt
fastapi
uvicorn
face_recognition
Pillow
python-multipart
scipy
numpy
qdrant-client
redis